requirements.txt
mamba-ssm==1.1.3
numpy==1.26.2
pytorch-lightning==2.1.2
rich==13.7.0
scikit-image==0.22.0
scikit-learn==1.3.2
scikit-learn-extra==0.3.0
scipy==1.11.4
seaborn==0.13.0
torchinfo==1.8.0
torch_geometric==2.4.0
tqdm==4.66.1
wandb==0.16.1
